Address review: a stripe index that cannot wrap, and gen in the column sweep
run_stripe and the backends' test-mode loops step a 64-bit index, so the
value after the last stripe is representable near the top of the domain
instead of wrapping back below shards; a run_stripe case at UINT32_MAX pins
it. The determinism class compares scav.geom.gen with the other columns --
every run is on a fresh chart, so its count is one like every other's -- and
the PRD's "every column" now holds literally. The reviewer's third point, a
lambda naming a const integral it does not capture, is not a defect: no
odr-use, no capture needed, and all six triples compile it.

+++ src/scav_thread.h
@@ -33,7 +33,11 @@ inline void run_stripe(uint32_t shards,
                        uint32_t first,
                        ShardFn fn,
                        void *ctx) {
-  for (uint32_t shard{ first }; shard < shards; shard += workers) { fn(ctx, shard); }
+  // 64-bit, so the index after the last stripe is representable near the top of
+  // the domain instead of wrapping back into it.
+  for (uint64_t shard{ first }; shard < shards; shard += workers) {
+    fn(ctx, static_cast<uint32_t>(shard));
+  }
 }
 
 }  // namespace scav
